ENG-11693: Add cluster IDs to DR producer and consumer stats.
- Add local cluster ID to the producer stats.

- Add remote cluster ID to the consumer stats.

+++ src/frontend/org/voltdb/ProducerDRGateway.java
@@ -68,7 +68,7 @@ public interface ProducerDRGateway {
 
     public abstract void updateCatalog(final CatalogContext catalog, final int listenPort);
 
-    public abstract int getDRClusterId();
+    public abstract byte getDRClusterId();
 
     public void cacheSnapshotRestoreTruncationPoint(Map<Integer, Long> sequenceNumbers);
